(server) Puntaje que suma al agarrar sorpresas o matar enemigos. El puntaje es global.

+++ programaServer/modelo/Metadata.h
@@ -18,7 +18,7 @@ public:
 
 	float escala;
 	int vidas[4];
-	int puntaje;
+	long int puntaje;
 	float posXCamara;
 	float posYCamara;
 	float tamanioXMundo;
